Fix basket serializer when tax is not known
When using the US strategy, tax isn't know until the shipping address is
entered. This causes BasketSerializer to throw an exception because access the
total_incl_tax, total_incl_tax_excl_discounts, or total_tax properties throw
TaxNotKnown when is_tax_known is false. This fixes the issue by detecting that
case and acknowledging in the API that tax isn't yet known.